```
benchmark                            old ns/op     new ns/op     delta
BenchmarkModelHasOneGopg-4           285887        248198        -13.18%
BenchmarkModelHasManyGopg-4          3093718       3043229       -1.63%
BenchmarkModelHasMany2ManyGopg-4     3378445       3265400       -3.35%

benchmark                            old allocs     new allocs     delta
BenchmarkModelHasOneGopg-4           1173           1173           +0.00%
BenchmarkModelHasManyGopg-4          8595           6596           -23.26%
BenchmarkModelHasMany2ManyGopg-4     9618           8619           -10.39%

benchmark                            old bytes     new bytes     delta
BenchmarkModelHasOneGopg-4           77572         77500         -0.09%
BenchmarkModelHasManyGopg-4          407260        390425        -4.13%
BenchmarkModelHasMany2ManyGopg-4     523806        508595        -2.90%
```